web: Redesign account panel, closes #141
